Move some models in the common.py.
