+++ huntflow_api_client/models/common.py
@@ -1,10 +1,19 @@
 import json
 from datetime import date, datetime
-from typing import AbstractSet, Any, Callable, Dict, Mapping, Optional, Union
+from typing import AbstractSet, Any, Callable, Dict, List, Mapping, Optional, Union
 
 from pydantic import AnyHttpUrl, BaseModel, EmailStr, Field, PositiveInt
 
-from huntflow_api_client.models.consts import EmailContactType, VacancyState
+from huntflow_api_client.models.consts import (
+    CalendarEventReminderMethod,
+    CalendarEventStatus,
+    CalendarEventType,
+    EmailContactType,
+    EventReminderMultiplier,
+    SurveyTypesEnum,
+    Transparency,
+    VacancyState,
+)
 
 IntStr = Union[int, str]
 AbstractSetIntStr = AbstractSet[IntStr]
@@ -185,4 +194,191 @@ class EmailFollowup(BaseModel):
         ...,
         gte=1,
         description="The number of days after which to send a followup if there is no response",
+    )
+
+
+class CalendarEventReminder(BaseModel):
+    multiplier: EventReminderMultiplier = Field(..., description="Reminder period")
+    value: int = Field(..., gte=0, lt=40320, description="Reminder value")
+    method: CalendarEventReminderMethod = Field(..., description="Reminder method")
+
+
+class CalendarEventAttendee(BaseModel):
+    member: Optional[PositiveInt] = Field(None, description="Coworker ID")
+    name: Optional[str] = Field(None, description="Attendee name", alias="displayName")
+    email: EmailStr = Field(..., description="Attendee email")
+
+    class Config:
+        allow_population_by_field_name = True
+
+
+class ApplicantLogEmail(BaseModel):
+    account_email: PositiveInt = Field(
+        ...,
+        description="Email account ID",
+    )
+    files: Optional[List[PositiveInt]] = Field(
+        None,
+        description="List of uploaded files ID",
+    )
+    followups: Optional[List[EmailFollowup]] = Field(
+        None,
+        description="List of email templates",
+    )
+    html: str = Field(..., description="Email content (HTML)")
+    email: EmailStr = Field(..., description="Recipient email address")
+    subject: str = Field(..., description="Email subject")
+    send_at: Optional[datetime] = Field(
+        None,
+        description="Date and time to send email. If not supplied, email will be sent immediately",
+    )
+    timezone: Optional[str] = Field(None, description="Time zone")
+    to: Optional[List[EmailRecipient]] = Field(
+        None,
+        description="List of additional recipients (cc/bcc)",
+    )
+    reply: Optional[PositiveInt] = Field(None, description="Reply email ID")
+
+
+class ApplicantLogIm(BaseModel):
+    account_im: PositiveInt = Field(..., description="Account IM ID")
+    receiver: str = Field(..., description="Username or phone of recipient")
+    body: str = Field(..., description="Message text")
+
+
+class ApplicantLogSms(BaseModel):
+    phone: str = Field(..., description="Phone number of recipient")
+    body: str = Field(..., description="Message text")
+
+
+class ApplicantOffer(BaseModel):
+    account_applicant_offer: PositiveInt = Field(..., description="Organization offer ID")
+    values: dict = Field(..., description="Organization offer schema")
+
+
+class ApplicantLogAccountInfo(BaseModel):
+    id: int = Field(..., description="ID of the user who created the log")
+    name: str = Field(..., description="Name of the user who created the log")
+
+
+class CalendarEventCreator(BaseModel):
+    name: Optional[str] = Field(
+        None,
+        description="Event creator name",
+    )
+    email: str = Field(..., description="Event creator email")
+    self: Optional[bool] = Field(
+        False,
+        description="Flag indicating that you are the creator of the event",
+    )
+
+
+class ApplicantLogCalendarEvent(BaseModel):
+    id: PositiveInt = Field(..., description="Calendar event ID")
+    name: Optional[str] = Field(
+        None,
+        description="Event name",
+    )
+    all_day: bool = Field(
+        ...,
+        description="Flag indicating that the event is scheduled for the whole day",
+    )
+    created: datetime = Field(..., description="Date and time of event creation")
+    creator: Optional[CalendarEventCreator] = Field(None, description="Event creator")
+    description: Optional[str] = Field(
+        None,
+        description="Event description",
+        example="Interview with John Doe",
+    )
+    timezone: Optional[str] = Field(None, description="Event time zone")
+    start: datetime = Field(..., description="Event start date and time")
+    end: datetime = Field(..., description="Event end date and time")
+    etag: Optional[str] = Field(None, description="Event Etag")
+    event_type: CalendarEventType = Field(..., description="Event type")
+    interview_type: Optional[int] = Field(None, description="Interview type ID")
+    calendar: PositiveInt = Field(..., description="Calendar ID")
+    vacancy: Optional[PositiveInt] = Field(
+        None,
+        description="Vacancy ID",
+    )
+    foreign: Optional[str] = Field(None, description="Foreign ID of event")
+    location: Optional[str] = Field(None, description="Event location")
+    attendees: List[CalendarEventAttendee] = Field(
+        [],
+        description="Event attendees (participants)",
+    )
+    reminders: List[CalendarEventReminder] = Field(
+        [],
+        description="List of reminders <a href=https://tools.ietf.org/html/rfc5545>RFC 5545</a>",
+    )
+    status: CalendarEventStatus = Field(..., description="Event status")
+    transparency: Transparency = Field(..., description="Event transparency (availability)")
+    recurrence: Optional[List] = None
+
+
+class ApplicantLogEmailResponse(BaseModel):
+    id: int = Field(..., description="Email ID")
+    created: datetime = Field(
+        ...,
+        description="Date and time of creating email",
+    )
+    subject: Optional[str] = Field(None, description="Email subject")
+    email_thread: Optional[int] = Field(None, description="Email thread ID")
+    account_email: Optional[int] = Field(
+        None,
+        description="Email account ID",
+    )
+    files: Optional[List[File]] = Field(None, description="List of uploaded files ID")
+    foreign: Optional[str] = Field(None, description="Foreign email ID")
+    timezone: str = Field(
+        ...,
+        description="Time zone",
+    )
+    html: Optional[str] = Field(
+        None,
+        description="Email content (HTML)",
+    )
+    from_email: Optional[str] = Field(
+        None,
+        description="Sender email address",
+    )
+    from_name: Optional[str] = Field(
+        None,
+        description="Sender name",
+    )
+    replyto: Optional[List[str]] = Field(
+        None,
+        description="List of email foreign IDs, to which a reply is send",
+    )
+    send_at: Optional[datetime] = Field(
+        None,
+        description="Date and time to send email",
+    )
+    to: Optional[List[EmailRecipient]] = Field(None, description="Recipients list")
+    state: Optional[str] = Field(None, description="Email state")
+
+
+class BaseSurveySchemaType(BaseModel):
+    id: int = Field(..., description="Survey ID")
+    name: str = Field(..., description="Survey name")
+    type: SurveyTypesEnum = Field(..., description="Survey type")
+    active: bool = Field(..., description="Is survey active?")
+    created: datetime = Field(..., description="Date and time of creating a survey")
+    updated: datetime = Field(..., description="Date and time of the last update of the survey")
+
+
+class SurveySchemaTypeQLogResponse(BaseSurveySchemaType):
+    title: Optional[str] = Field(..., description="Survey title")
+
+
+class ApplicantLogSurveyQuestionary(BaseModel):
+    id: int = Field(..., description="Survey questionary ID")
+    survey: SurveySchemaTypeQLogResponse = Field(..., description="Survey schema")
+    survey_answer_id: Optional[int] = Field(
+        None,
+        description="Survey questionary answer ID",
+    )
+    created: datetime = Field(
+        ...,
+        description="Date and time of creating an survey questionary",
     )
