+++ huntflow_api_client/models/request/applicants.py
@@ -1,20 +1,15 @@
 from datetime import date, datetime
 from typing import List, Literal, Optional
 
-from pydantic import BaseModel, EmailStr, Field, PositiveInt
+from pydantic import BaseModel, Field, PositiveInt
 
 from huntflow_api_client.models.common import (
     Applicant,
-    EmailFollowup,
-    EmailRecipient,
+    CalendarEventAttendee,
+    CalendarEventReminder,
     JsonRequestModel,
 )
-from huntflow_api_client.models.consts import (
-    CalendarEventReminderMethod,
-    CalendarEventType,
-    EventReminderMultiplier,
-    Transparency,
-)
+from huntflow_api_client.models.consts import CalendarEventType, Transparency
 
 
 class ApplicantResumeData(BaseModel):
@@ -57,65 +52,6 @@ class ApplicantUpdateRequest(Applicant, JsonRequestModel):
         max_items=1,
         description="List of applicant's social accounts",
     )
-
-
-class CalendarEventReminder(BaseModel):
-    multiplier: EventReminderMultiplier = Field(..., description="Reminder period")
-    value: int = Field(..., gte=0, lt=40320, description="Reminder value")
-    method: CalendarEventReminderMethod = Field(..., description="Reminder method")
-
-
-class CalendarEventAttendee(BaseModel):
-    member: Optional[PositiveInt] = Field(None, description="Coworker ID")
-    name: Optional[str] = Field(None, description="Attendee name", alias="displayName")
-    email: EmailStr = Field(..., description="Attendee email")
-
-    class Config:
-        allow_population_by_field_name = True
-
-
-class ApplicantLogEmail(BaseModel):
-    account_email: PositiveInt = Field(
-        ...,
-        description="Email account ID",
-    )
-    files: Optional[List[PositiveInt]] = Field(
-        None,
-        description="List of uploaded files ID",
-    )
-    followups: Optional[List[EmailFollowup]] = Field(
-        None,
-        description="List of email templates",
-    )
-    html: str = Field(..., description="Email content (HTML)")
-    email: EmailStr = Field(..., description="Recipient email address")
-    subject: str = Field(..., description="Email subject")
-    send_at: Optional[datetime] = Field(
-        None,
-        description="Date and time to send email. If not supplied, email will be sent immediately",
-    )
-    timezone: Optional[str] = Field(None, description="Time zone")
-    to: Optional[List[EmailRecipient]] = Field(
-        None,
-        description="List of additional recipients (cc/bcc)",
-    )
-    reply: Optional[PositiveInt] = Field(None, description="Reply email ID")
-
-
-class ApplicantLogIm(BaseModel):
-    account_im: PositiveInt = Field(..., description="Account IM ID")
-    receiver: str = Field(..., description="Username or phone of recipient")
-    body: str = Field(..., description="Message text")
-
-
-class ApplicantLogSms(BaseModel):
-    phone: str = Field(..., description="Phone number of recipient")
-    body: str = Field(..., description="Message text")
-
-
-class ApplicantOffer(BaseModel):
-    account_applicant_offer: PositiveInt = Field(..., description="Organization offer ID")
-    values: dict = Field(..., description="Organization offer schema")
 
 
 class ApplicantEvent(BaseModel):
