+++ huntflow_api_client/models/response/applicants.py
@@ -3,16 +3,9 @@ from typing import List, Optional, Union
 
 from pydantic import BaseModel, EmailStr, Field, PositiveInt
 
-from huntflow_api_client.models.common import Applicant, File, PaginatedResponse
+from huntflow_api_client.models.common import Applicant, PaginatedResponse
 from huntflow_api_client.models.consts import AgreementState as AgreementStateEnum
-from huntflow_api_client.models.consts import (
-    CalendarEventReminderMethod,
-    CalendarEventStatus,
-    CalendarEventType,
-    EmailContactType,
-    Transparency,
-)
-from huntflow_api_client.models.response.survey import SurveySchemaTypeQLogResponse
+from huntflow_api_client.models.consts import CalendarEventReminderMethod, CalendarEventStatus
 
 
 class ApplicantTag(BaseModel):
@@ -144,74 +137,6 @@ class ApplicantSearchByCursorResponse(BaseModel):
     next_page_cursor: Optional[str] = Field(None, description="Next page cursor")
 
 
-class EmailRecipient(BaseModel):
-    type: Optional[EmailContactType] = Field(None, description="Type of the email contact")
-    name: Optional[str] = Field(
-        None,
-        description="Name of email recipient",
-    )
-    email: str = Field(..., description="Email address")
-
-
-class ApplicantLogEmailResponse(BaseModel):
-    id: int = Field(..., description="Email ID")
-    created: datetime = Field(
-        ...,
-        description="Date and time of creating email",
-    )
-    subject: Optional[str] = Field(None, description="Email subject")
-    email_thread: Optional[int] = Field(None, description="Email thread ID")
-    account_email: Optional[int] = Field(
-        None,
-        description="Email account ID",
-    )
-    files: Optional[List[File]] = Field(None, description="List of uploaded files ID")
-    foreign: Optional[str] = Field(None, description="Foreign email ID")
-    timezone: str = Field(
-        ...,
-        description="Time zone",
-    )
-    html: Optional[str] = Field(
-        None,
-        description="Email content (HTML)",
-    )
-    from_email: Optional[str] = Field(
-        None,
-        description="Sender email address",
-    )
-    from_name: Optional[str] = Field(
-        None,
-        description="Sender name",
-    )
-    replyto: Optional[List[str]] = Field(
-        None,
-        description="List of email foreign IDs, to which a reply is send",
-    )
-    send_at: Optional[datetime] = Field(
-        None,
-        description="Date and time to send email",
-    )
-    to: Optional[List[EmailRecipient]] = Field(None, description="Recipients list")
-    state: Optional[str] = Field(None, description="Email state")
-
-
-class ApplicantLogAccountInfo(BaseModel):
-    id: int = Field(..., description="ID of the user who created the log")
-    name: str = Field(..., description="Name of the user who created the log")
-
-
-class CalendarEventCreator(BaseModel):
-    name: Optional[str] = Field(
-        None,
-        description="Event creator name",
-    )
-    email: str = Field(..., description="Event creator email")
-    self: Optional[bool] = Field(
-        False,
-        description="Flag indicating that you are the creator of the event",
-    )
-
-
 class CalendarEventAttendee(BaseModel):
     member: Optional[int] = Field(None, description="Coworker ID")
     name: Optional[str] = Field(
@@ -234,59 +159,3 @@ class CalendarEventReminder(BaseModel):
         description="Reminder method",
     )
     minutes: int = Field(..., description="How many minutes in advance to remind about the event")
-
-
-class ApplicantLogCalendarEvent(BaseModel):
-    id: PositiveInt = Field(..., description="Calendar event ID")
-    name: Optional[str] = Field(
-        None,
-        description="Event name",
-    )
-    all_day: bool = Field(
-        ...,
-        description="Flag indicating that the event is scheduled for the whole day",
-    )
-    created: datetime = Field(..., description="Date and time of event creation")
-    creator: Optional[CalendarEventCreator] = Field(None, description="Event creator")
-    description: Optional[str] = Field(
-        None,
-        description="Event description",
-        example="Interview with John Doe",
-    )
-    timezone: Optional[str] = Field(None, description="Event time zone")
-    start: datetime = Field(..., description="Event start date and time")
-    end: datetime = Field(..., description="Event end date and time")
-    etag: Optional[str] = Field(None, description="Event Etag")
-    event_type: CalendarEventType = Field(..., description="Event type")
-    interview_type: Optional[int] = Field(None, description="Interview type ID")
-    calendar: PositiveInt = Field(..., description="Calendar ID")
-    vacancy: Optional[PositiveInt] = Field(
-        None,
-        description="Vacancy ID",
-    )
-    foreign: Optional[str] = Field(None, description="Foreign ID of event")
-    location: Optional[str] = Field(None, description="Event location")
-    attendees: List[CalendarEventAttendee] = Field(
-        [],
-        description="Event attendees (participants)",
-    )
-    reminders: List[CalendarEventReminder] = Field(
-        [],
-        description="List of reminders <a href=https://tools.ietf.org/html/rfc5545>RFC 5545</a>",
-    )
-    status: CalendarEventStatus = Field(..., description="Event status")
-    transparency: Transparency = Field(..., description="Event transparency (availability)")
-    recurrence: Optional[List] = None
-
-
-class ApplicantLogSurveyQuestionary(BaseModel):
-    id: int = Field(..., description="Survey questionary ID")
-    survey: SurveySchemaTypeQLogResponse = Field(..., description="Survey schema")
-    survey_answer_id: Optional[int] = Field(
-        None,
-        description="Survey questionary answer ID",
-    )
-    created: datetime = Field(
-        ...,
-        description="Date and time of creating an survey questionary",
-    )
